# Heap-free hashing