There is no good reason why hashing a block of data should require heap allocation or even a non-trivial dynamic buffer.

Unfortunately, the murmur library makes this a bit difficult. Thankfully, we can just roll our own thin functions. It is possible to do while preserving exactly the same hash values. That's what this PR does.

The goal of this PR is not to improve speed. Of course, it does so... On my machine...

Before:
```
BenchmarkEstimated
BenchmarkEstimated-8                    1000000000               0.1333 ns/op          0 B/op          0 allocs/op
BenchmarkSeparateTestAndAdd
BenchmarkSeparateTestAndAdd-8            2706183               508.2 ns/op           194 B/op          4 allocs/op
BenchmarkCombinedTestAndAdd
BenchmarkCombinedTestAndAdd-8            3312842               437.6 ns/op            97 B/op          2 allocs/op
PASS
```

After...

```
BenchmarkEstimated-8                    1000000000               0.05285 ns/op         0 B/op          0 allocs/op
BenchmarkSeparateTestAndAdd
BenchmarkSeparateTestAndAdd-8            6219651               451.9 ns/op             0 B/op          0 allocs/op
BenchmarkCombinedTestAndAdd
BenchmarkCombinedTestAndAdd-8            5010103               393.2 ns/op             0 B/op          0 allocs/op
PASS
```

So we see about a 10% boost in performance in this instance on my system. Results will vary.

The important part is that it does away entirely with heap allocation.

The PR includes tests to ensure that the hash values are preserved.

Note that if we break backward compatibility (with exact hash value reproduction), there are many great optimization opportunities. But that should be addressed in other PRs.

Related to PRs https://github.com/willf/bloom/pull/51 and https://github.com/willf/bloom/pull/57