@lemire 2021-04-25
You are right, I removed `;`. I think it should be possible to do much better in pure Go by optimizing the hash function further. However, removing the heap allocation should help tremendously in practice.