@lemire 2021-04-29
@moredure has reviewed and has approved.

Merging.